requirements.txt
requests==2.31.0
tmdbv3api==1.7.7
anthropic==0.40.0
pysrt==1.1.2
python-dotenv==1.0.0
flask==3.0.0